converted thunks to hooks
